portfolios-apis/okrs-api-main/pyproject.toml
[tool.poetry]
name = "okrs-api"
version = "6.0.1"
description = "The API for the OKRs project"
authors = ["author_name <platforma@planview.com>"]
keywords = [
  "openapi",
  "rest",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Web Environment",
  "Intended Audience :: End Users/Desktop",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Microsoft :: Windows",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
  "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
]

[[tool.poetry.source]]
name = "platforma"
url = "https://platforma-486329218230.d.codeartifact.us-west-2.amazonaws.com/pypi/platforma/simple/"
secondary = true

[tool.poetry.dependencies]
python = "^3.9"
OpenAlchemy =  "^2.0.0"
PyJWT = "^2.4.0"
SQLAlchemy = "^1.4.2"
aiohttp = "3.10.11"
aiohttp-jinja2 = "^1.4.2"
aiohttp_cors = "^0.7.0"
alembic = "^1.8.1"
connexion = "^2.7.0"
loguru = "^0.6.0"
psycopg2 = "2.9.1"
markupsafe = "2.1.3"
pydantic = "^1.6.1"
pyyaml = "^6.0.1"
sentry-sdk = "^2.8.0"
swagger-ui-bundle = "^0.0.8"
tenacity = "^6.2.0"
uvloop = "^0.14.0"
inflection = "^0.5.1"
gunicorn = "^22.0.0"
boto3 = "^1.20.23"
pubnub = ">=7.0.1"
Werkzeug = "~3.0.6"
flask = "=2.2.5"
requests = "2.32.2"
certifi = "2024.07.04"
urllib3 = "1.26.19"

[tool.poetry.dev-dependencies]
Faker = "^4.1.3"
Sphinx = "^7.2.6"
aiohttp-theme = "^0.1.6"
black = "^21.10b0"
bpython = "^0.21"
coveralls = "^3.0.0"
faker-planview = "^0.4.0"
flake8 = "^3.7.9"
invoke = "^1.4.1"
jinja2 = "^3.1.6"
platforma-invocations = "^1.3.2"
pydocstyle = "^5.0.2"
pytest = "^6.1.1"
pytest-aiohttp = "^0.3.0"
pytest-cov = "^2.8.1"
pytest-mock = "^3.1.0"
pytest-rerunfailures = "^9.1.1"
pytest-socket = "^0.3.5"
pytest-vcr = "^1.0.2"
pytest-xdist = "^2.1.0"
recommonmark = "^0.7.1"
typer = "0.4.0"
click = "8.0"
pylint = "<2.7.0"
sphinx-copybutton = "^0.3.1"
sphinx-autobuild = "^2020.9.1"
myst-parser = "^2.0.0"
furo = "^2024.01.29"
vcrpy = {git = "https://github.com/vprakashplanview/vcrpy.git"}
mock-alchemy = "^0.2.1"
factory-boy = "^3.2.0"
pytest-factoryboy = "^2.1.0"
tomlkit = "^0.7.2"
pytest-dotenv = "^0.5.2"

[tool.poetry.scripts]
okrs-api = "okrs_api.main:main"
account-export = "okrs_api.account_export:main"
account-import = "okrs_api.account_import:main"
account-migrate = "okrs_api.account_migrate:main"

[tool.black]
exclude = '''
(
  \.git
  |.venv
  |api\/models_autogenerated\.py
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

